Improve typing (electron-app)
