+++ electron-app/build-templates/grapevne-build.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+docker build -t grapevne-build .
+
+# -t = allocate a pseudo-TTY (passes syntax highlights from the terminal)
+docker run -t grapevne-build
